# docs/swagger.yaml
definitions:
  genprotos.Candidate:
    properties:
      date:
        type: string
      election:
        $ref: '#/definitions/genprotos.Election'
      id:
        type: string
      party:
        $ref: '#/definitions/genprotos.Party'
      public:
        $ref: '#/definitions/genprotos.Public'
    type: object
  genprotos.Election:
    properties:
      date:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  genprotos.GetAllCandidate:
    properties:
      Candidates:
        items:
          $ref: '#/definitions/genprotos.Candidate'
        type: array
    type: object
  genprotos.GetAllElection:
    properties:
      elections:
        items:
          $ref: '#/definitions/genprotos.Election'
        type: array
    type: object
  genprotos.GetAllParty:
    properties:
      partys:
        items:
          $ref: '#/definitions/genprotos.Party'
        type: array
    type: object
  genprotos.GetAllPublic:
    properties:
      publics:
        items:
          $ref: '#/definitions/genprotos.Public'
        type: array
    type: object
  genprotos.GetAllPublicVote:
    properties:
      publicVotes:
        items:
          $ref: '#/definitions/genprotos.PublicVote'
        type: array
    type: object
  genprotos.Party:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      open_date:
        type: string
      slogan:
        type: string
    type: object
  genprotos.Public:
    properties:
      birthday:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      nation:
        type: string
      party:
        $ref: '#/definitions/genprotos.Party'
    type: object
  genprotos.PublicVote:
    properties:
      candidate:
        $ref: '#/definitions/genprotos.Candidate'
      election:
        $ref: '#/definitions/genprotos.Election'
      id:
        type: string
      public:
        $ref: '#/definitions/genprotos.Public'
    type: object
info:
  contact: {}
paths:
  /PublicVote/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: GetById page
      parameters:
      - description: PublicVote ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetById Successful
          schema:
            $ref: '#/definitions/genprotos.PublicVote'
        "401":
          description: Error while GetByIdd
          schema:
            type: string
      summary: GetById PublicVote
      tags:
      - PublicVote
  /candidate/create:
    post:
      consumes:
      - application/json
      description: Create page
      parameters:
      - description: Create
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genprotos.Candidate'
      produces:
      - application/json
      responses:
        "200":
          description: Create Successful
          schema:
            type: string
        "401":
          description: Error while Created
          schema:
            type: string
      summary: Create Candidate
      tags:
      - Candidate
  /candidate/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete page
      parameters:
      - description: Candidate ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete Successful
          schema:
            type: string
        "401":
          description: Error while Deleted
          schema:
            type: string
      summary: Delete Candidate
      tags:
      - Candidate
  /candidate/getall:
    get:
      consumes:
      - application/json
      description: GetAll page
      produces:
      - application/json
      responses:
        "200":
          description: GetAll Successful
          schema:
            $ref: '#/definitions/genprotos.GetAllCandidate'
        "401":
          description: Error while GetAlld
          schema:
            type: string
      summary: GetAll Candidate
      tags:
      - Candidate
  /candidate/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: GetById page
      parameters:
      - description: Candidate ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetById Successful
          schema:
            $ref: '#/definitions/genprotos.Candidate'
        "401":
          description: Error while GetByIdd
          schema:
            type: string
      summary: GetById Candidate
      tags:
      - Candidate
  /candidate/update/{id}:
    put:
      consumes:
      - application/json
      description: Update page
      parameters:
      - description: Candidate ID
        in: path
        name: id
        required: true
        type: string
      - description: Update
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genprotos.Candidate'
      produces:
      - application/json
      responses:
        "200":
          description: Update Successful
          schema:
            type: string
        "401":
          description: Error while created
          schema:
            type: string
      summary: Update Candidate
      tags:
      - Candidate
  /election/create:
    post:
      consumes:
      - application/json
      description: Create page
      parameters:
      - description: Create
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genprotos.Election'
      produces:
      - application/json
      responses:
        "200":
          description: Create Successful
          schema:
            type: string
        "401":
          description: Error while Created
          schema:
            type: string
      summary: Create Election
      tags:
      - Election
  /election/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete page
      parameters:
      - description: Election ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete Successful
          schema:
            type: string
        "401":
          description: Error while Deleted
          schema:
            type: string
      summary: Delete Election
      tags:
      - Election
  /election/getall:
    get:
      consumes:
      - application/json
      description: GetAll page
      produces:
      - application/json
      responses:
        "200":
          description: GetAll Successful
          schema:
            $ref: '#/definitions/genprotos.GetAllElection'
        "401":
          description: Error while GetAlld
          schema:
            type: string
      summary: GetAll Election
      tags:
      - Election
  /election/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: GetById page
      parameters:
      - description: Election ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetById Successful
          schema:
            $ref: '#/definitions/genprotos.Election'
        "401":
          description: Error while GetByIdd
          schema:
            type: string
      summary: GetById Election
      tags:
      - Election
  /election/update/{id}:
    put:
      consumes:
      - application/json
      description: Update page
      parameters:
      - description: Election ID
        in: path
        name: id
        required: true
        type: string
      - description: Update
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genprotos.Election'
      produces:
      - application/json
      responses:
        "200":
          description: Update Successful
          schema:
            type: string
        "401":
          description: Error while created
          schema:
            type: string
      summary: Update Election
      tags:
      - Election
  /party/create:
    post:
      consumes:
      - application/json
      description: Create page
      parameters:
      - description: Create
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genprotos.Party'
      produces:
      - application/json
      responses:
        "200":
          description: Create Successful
          schema:
            type: string
        "401":
          description: Error while Created
          schema:
            type: string
      summary: Create Party
      tags:
      - Party
  /party/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete page
      parameters:
      - description: Party ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete Successful
          schema:
            type: string
        "401":
          description: Error while Deleted
          schema:
            type: string
      summary: Delete Party
      tags:
      - Party
  /party/getall:
    get:
      consumes:
      - application/json
      description: GetAll page
      produces:
      - application/json
      responses:
        "200":
          description: GetAll Successful
          schema:
            $ref: '#/definitions/genprotos.GetAllParty'
        "401":
          description: Error while GetAlld
          schema:
            type: string
      summary: GetAll Party
      tags:
      - Party
  /party/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: GetById page
      parameters:
      - description: Party ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetById Successful
          schema:
            $ref: '#/definitions/genprotos.Party'
        "401":
          description: Error while GetByIdd
          schema:
            type: string
      summary: GetById Party
      tags:
      - Party
  /party/update/{id}:
    put:
      consumes:
      - application/json
      description: Update page
      parameters:
      - description: Party ID
        in: path
        name: id
        required: true
        type: string
      - description: Update
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genprotos.Party'
      produces:
      - application/json
      responses:
        "200":
          description: Update Successful
          schema:
            type: string
        "401":
          description: Error while created
          schema:
            type: string
      summary: Update Party
      tags:
      - Party
  /public/cheak/{id}:
    get:
      consumes:
      - application/json
      description: Cheak page
      parameters:
      - description: Public ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cheak Successful
          schema:
            type: string
        "401":
          description: Error while GetByIdd
          schema:
            type: string
      summary: Cheak Public
      tags:
      - Public
  /public/create:
    post:
      consumes:
      - application/json
      description: Create page
      parameters:
      - description: Create
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genprotos.Public'
      produces:
      - application/json
      responses:
        "200":
          description: Create Successful
          schema:
            type: string
        "401":
          description: Error while Created
          schema:
            type: string
      summary: Create Public
      tags:
      - Public
  /public/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete page
      parameters:
      - description: Public ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete Successful
          schema:
            type: string
        "401":
          description: Error while Deleted
          schema:
            type: string
      summary: Delete Public
      tags:
      - Public
  /public/getall:
    get:
      consumes:
      - application/json
      description: GetAll page
      produces:
      - application/json
      responses:
        "200":
          description: GetAll Successful
          schema:
            $ref: '#/definitions/genprotos.GetAllPublic'
        "401":
          description: Error while GetAlld
          schema:
            type: string
      summary: GetAll Public
      tags:
      - Public
  /public/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: GetById page
      parameters:
      - description: Public ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetById Successful
          schema:
            $ref: '#/definitions/genprotos.Public'
        "401":
          description: Error while GetByIdd
          schema:
            type: string
      summary: GetById Public
      tags:
      - Public
  /public/update/{id}:
    put:
      consumes:
      - application/json
      description: Update page
      parameters:
      - description: Public ID
        in: path
        name: id
        required: true
        type: string
      - description: Update
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genprotos.Public'
      produces:
      - application/json
      responses:
        "200":
          description: Update Successful
          schema:
            type: string
        "401":
          description: Error while created
          schema:
            type: string
      summary: Update Public
      tags:
      - Public
  /public_vote/create:
    post:
      consumes:
      - application/json
      description: Create page
      parameters:
      - description: Create
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genprotos.PublicVote'
      produces:
      - application/json
      responses:
        "200":
          description: Create Successful
          schema:
            type: string
        "401":
          description: Error while Created
          schema:
            type: string
      summary: Create Public_Vote
      tags:
      - PublicVote
  /publicvote/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete page
      parameters:
      - description: PublicVote ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete Successful
          schema:
            type: string
        "401":
          description: Error while Deleted
          schema:
            type: string
      summary: Delete PublicVote
      tags:
      - PublicVote
  /publicvote/getall:
    get:
      consumes:
      - application/json
      description: GetAll page
      produces:
      - application/json
      responses:
        "200":
          description: GetAll Successful
          schema:
            $ref: '#/definitions/genprotos.GetAllPublicVote'
        "401":
          description: Error while GetAlld
          schema:
            type: string
      summary: GetAll PublicVote
      tags:
      - PublicVote
  /publicvote/update/{id}:
    put:
      consumes:
      - application/json
      description: Update page
      parameters:
      - description: PublicVOTE ID
        in: path
        name: id
        required: true
        type: string
      - description: Update
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genprotos.PublicVote'
      produces:
      - application/json
      responses:
        "200":
          description: Update Successful
          schema:
            type: string
        "401":
          description: Error while created
          schema:
            type: string
      summary: Update PublicVOTE
      tags:
      - PublicVote
swagger: "2.0"